[SPARK-46418][PS][TESTS] Reorganize `ReshapeTests`
### What changes were proposed in this pull request?
break `ReshapeTests` into multiple small tests

### Why are the changes needed?
1, it parity test is slow, sometimes takes >5 mins;
2, to be consistent with pandas' test https://github.com/pandas-dev/pandas/tree/main/pandas/tests/reshape

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44365 from zhengruifeng/ps_test_reshape.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
