+++ dev/sparktestsupport/modules.py
@@ -754,7 +754,12 @@ pyspark_pandas = Module(
         "pyspark.pandas.tests.resample.test_missing",
         "pyspark.pandas.tests.resample.test_series",
         "pyspark.pandas.tests.resample.test_timezone",
-        "pyspark.pandas.tests.test_reshape",
+        "pyspark.pandas.tests.reshape.test_get_dummies",
+        "pyspark.pandas.tests.reshape.test_get_dummies_kwargs",
+        "pyspark.pandas.tests.reshape.test_get_dummies_multiindex",
+        "pyspark.pandas.tests.reshape.test_get_dummies_object",
+        "pyspark.pandas.tests.reshape.test_get_dummies_prefix",
+        "pyspark.pandas.tests.reshape.test_merge_asof",
         "pyspark.pandas.tests.window.test_rolling",
         "pyspark.pandas.tests.window.test_rolling_adv",
         "pyspark.pandas.tests.window.test_rolling_count",
@@ -1104,7 +1109,12 @@ pyspark_pandas_connect_part1 = Module(
         "pyspark.pandas.tests.connect.series.test_parity_sort",
         "pyspark.pandas.tests.connect.series.test_parity_stat",
         "pyspark.pandas.tests.connect.data_type_ops.test_parity_num_arithmetic",
-        "pyspark.pandas.tests.connect.test_parity_reshape",
+        "pyspark.pandas.tests.connect.reshape.test_parity_get_dummies",
+        "pyspark.pandas.tests.connect.reshape.test_parity_get_dummies_kwargs",
+        "pyspark.pandas.tests.connect.reshape.test_parity_get_dummies_multiindex",
+        "pyspark.pandas.tests.connect.reshape.test_parity_get_dummies_object",
+        "pyspark.pandas.tests.connect.reshape.test_parity_get_dummies_prefix",
+        "pyspark.pandas.tests.connect.reshape.test_parity_merge_asof",
         "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames_groupby_expanding",
     ],
     excluded_python_implementations=[
