+++ python/pyspark/pandas/tests/connect/reshape/test_parity_get_dummies_kwargs.py
@@ -16,17 +16,21 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_reshape import ReshapeTestsMixin
+from pyspark.pandas.tests.reshape.test_get_dummies_kwargs import GetDummiesKWArgsMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 
 
-class ReshapeParityTests(ReshapeTestsMixin, PandasOnSparkTestUtils, ReusedConnectTestCase):
+class GetDummiesKWArgsParityTests(
+    GetDummiesKWArgsMixin,
+    PandasOnSparkTestUtils,
+    ReusedConnectTestCase,
+):
     pass
 
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_reshape import *  # noqa: F401
+    from pyspark.pandas.tests.connect.reshape.test_parity_get_dummies_kwargs import *  # noqa
 
     try:
         import xmlrunner  # type: ignore[import]
